SDK-36: Split tests to multiple files. Updated validation logic for required params.
